# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminRoleReplaceRequest": {
        "description": "`PUT /admin/users/{id}/roles` —— 全量替换。",
        "properties": {
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleCode"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "roles"
        ],
        "title": "AdminRoleReplaceRequest",
        "type": "object"
      },
...
      "AdminUserCreateRequest": {
        "properties": {
          "display_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "must_change_password": {
            "default": true,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "留空则服务端生成随机密码，并在响应的 generated_password 里回显一次",
            "title": "Password"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleCode"
            },
            "title": "Roles",
            "type": "array"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "display_name"
        ],
        "title": "AdminUserCreateRequest",
        "type": "object"
      },
      "AdminUserCreateResponse": {
        "properties": {
          "generated_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Password"
          },
          "user": {
            "$ref": "#/components/schemas/AdminUserItem"
          }
        },
        "required": [
          "user"
        ],
        "title": "AdminUserCreateResponse",
        "type": "object"
      },
      "AdminUserItem": {
        "properties": {
          "active_token_count": {
            "default": 0,
            "title": "Active Token Count",
            "type": "integer"
          },
          "auth_source": {
            "title": "Auth Source",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "failed_login_count": {
            "default": 0,
            "title": "Failed Login Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "last_login_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login Ip"
          },
          "locked_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Until"
          },
          "must_change_password": {
            "default": false,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "tool_count": {
            "default": 0,
            "title": "Tool Count",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "used_bytes": {
            "default": 0,
            "title": "Used Bytes",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "status",
          "auth_source"
        ],
        "title": "AdminUserItem",
        "type": "object"
      },
      "AdminUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "AdminUserListResponse",
        "type": "object"
      },
      "AdminUserUpdateRequest": {
        "description": "PATCH —— 只更新传入的字段。`username` **不可改**（FR-IAM-01）。",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AdminUserUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "title": "ResetPasswordRequest",
        "type": "object"
      },
      "ResetPasswordResponse": {
        "properties": {
          "generated_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated Password"
          },
          "must_change_password": {
            "default": true,
            "title": "Must Change Password",
            "type": "boolean"
          },
          "revoked_sessions": {
            "default": 0,
            "title": "Revoked Sessions",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "ResetPasswordResponse",
        "type": "object"
      },
      "RoleCode": {
        "description": "docs/02 §3.2 `roles.code` —— 四角色固定集合（docs/01 §3.1）",
        "enum": [
          "superadmin",
          "approver",
          "user",
          "viewer"
        ],
        "title": "RoleCode",
        "type": "string"
      },
...
      "UserStatus": {
        "description": "docs/02 §3.1 `users.status`",
        "enum": [
          "active",
          "disabled"
        ],
        "title": "UserStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "selftool — 内网工具 / Skill 共享平台后端",
    "title": "selftool API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "description": "FR-IAM-04：模糊搜索 + 角色筛选 + 状态筛选 + 分页。",
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "description": "按用户名/显示名模糊搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按用户名/显示名模糊搜索",
              "title": "Q"
            }
          },
          {
            "description": "角色筛选，可重复",
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "角色筛选，可重复",
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "用户列表",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "FR-IAM-01。密码留空时生成随机密码并**只回显一次**。\n\n注意：`generated_password` 绝不出现在日志里。",
        "operationId": "create_user_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建用户",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "用户详情",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "description": "FR-IAM-02。禁用会连带吊销该用户的全部 refresh token 与 API Token。",
        "operationId": "update_user_api_v1_admin_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "编辑用户",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/reset-password": {
      "post": {
        "description": "FR-IAM-03：重置后强制改密，并吊销该用户全部会话。",
        "operationId": "reset_password_api_v1_admin_users__user_id__reset_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "重置密码",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/revoke-sessions": {
      "post": {
        "description": "FR-AUTH-12。",
        "operationId": "revoke_sessions_api_v1_admin_users__user_id__revoke_sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Revoke Sessions Api V1 Admin Users  User Id  Revoke Sessions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "强制下线（吊销全部会话）",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/roles": {
      "put": {
        "description": "FR-IAM-02。降级最后一个超管 → 409 LAST_SUPERADMIN；降级自己 → 被拒。",
        "operationId": "replace_roles_api_v1_admin_users__user_id__roles_put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRoleReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "全量替换角色",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}